\name{lab1}
\alias{lab1}
  
\title{
    Exercise 1 in advanced modeling in r
  }
\description{
    function named lab1 which produces regression coefficients done as an exercise for advanced R modeling.
  }
\usage{
 lab1(x,y)
   
  }
  
\arguments{

\item{x}{
value of x
    }
\item{y}{
      value of y
    }
    }
    
\author{
      Niels Kruse, \email{niels.kruse@uzh.ch}
    }
    

    
    \keyword{ regression, lm }