\name{lab2}
\alias{lab2}

  
\title{
    Exercise 2 in advanced modeling in r
  }
\description{
    function named lab2 which produces log likelihoood  scores done as an exercise for advanced R modeling.
  }
\usage{
    lab2(theta,y,X,Z)
    
  }
  
\arguments{

\item{theta}{
value of theta}
\item{Z}{
value of Z}
\item{X}{
value of X}

\item{y}{
      value of y
    }
    }
    
\author{
      Niels Kruse, \email{niels.kruse@uzh.ch}
    }
    

    
    \keyword{ regression, mls, log }